+++ src/components/fr-generator/widgets/antd/table.js
@@ -1,3 +1,4 @@
 
-import { Table } from 'antd';
+import { Table} from 'antd';
 import tableHoc from '../../components/tableHoc';
+export default tableHoc(Table);
